Revert "Move `test_result_post_process` out of execution.py"
This reverts commit 4e0b9a09d6932ca3043285bfa7a07723152300d8.
