+++ tests/covalent_tests/results_manager_tests/results_test.py
@@ -470,75 +470,3 @@ def test_initialize_result_object(mocker):
 
     mock_result_instance.persist.assert_called_with(electron_id=5)
     assert sub_result_object._root_dispatch_id == result_object.dispatch_id
-
-
-def test_result_post_process(mocker, test_db):
-    """Test client-side post-processing of results."""
-
-    import covalent as ct
-
-    @ct.electron
-    def construct_cu_slab(x):
-        return x
-
-    @ct.electron
-    def compute_system_energy(x):
-        return x
-
-    @ct.electron
-    def construct_n_molecule(x):
-        return x
-
-    @ct.electron
-    def get_relaxed_slab(x):
-        return x
-
-    @ct.lattice
-    def compute_energy():
-        N2 = construct_n_molecule(1)
-        e_N2 = compute_system_energy(N2)
-
-        slab = construct_cu_slab(2)
-        e_slab = compute_system_energy(slab)
-
-        relaxed_slab = get_relaxed_slab(3)
-        e_relaxed_slab = compute_system_energy(relaxed_slab)
-
-        return (N2, e_N2, slab, e_slab, relaxed_slab, e_relaxed_slab)
-
-    compute_energy.build_graph()
-
-    compute_energy = LatticeClass.deserialize_from_json(compute_energy.serialize_to_json())
-
-    node_outputs = {
-        "construct_n_molecule(0)": 1,
-        ":parameter:1(1)": 1,
-        "compute_system_energy(2)": 1,
-        "construct_cu_slab(3)": 2,
-        ":parameter:2(4)": 2,
-        "compute_system_energy(5)": 2,
-        "get_relaxed_slab(6)": 3,
-        ":parameter:3(7)": 3,
-        "compute_system_energy(8)": 3,
-    }
-
-    encoded_node_outputs = {
-        k: ct.TransportableObject.make_transportable(v) for k, v in node_outputs.items()
-    }
-
-    res = Result(compute_energy, compute_energy.metadata["results_dir"])
-    res._initialize_nodes()
-
-    for i, v in enumerate(encoded_node_outputs.values()):
-        compute_energy.transport_graph.set_node_value(i, "output", v)
-
-    res._status = Result.PENDING_POSTPROCESSING
-    res._dispatch_id = "MOCK"
-    res._root_dispatch_id = "MOCK"
-
-    mocker.patch("covalent._data_store.datastore.DataStore.factory", return_value=test_db)
-    res.persist()
-
-    execution_result = res.post_process()
-
-    assert execution_result == compute_energy()
